#169 Finished weltwaerts scraper
