+++ etl/data_enhancement/domain/weltwaerts.py
@@ -1,0 +1,2 @@
+def run(data):
+    pass
