+++ etl/data_extraction/scraper/weltwaerts.py
@@ -1,71 +1,90 @@
-import scrapy
-from scrapy.crawler import CrawlerProcess
+from data_extraction.Scraper import Scraper
 
 
+class WeltwaertsScraper(Scraper):
+    base_url = 'https://www.weltwaerts.de'
+    debug = True
 
-class crawler(scrapy.Spider):
-    name = 'crawler'
-    allowed_domains = ['weltwaerts.de']
-    start_urls = ['https://www.weltwaerts.de/de/ep-ergebnis.html']
+    #   Handles the soupified response of a detail page in the predefined way and returns it
+    def parse(self, response, url):
+        import re
 
+        param_box = response.find('div', {'class': 'parameter__box'})
+        content = response.find('div', {'class': 'mod_epdetail__content-container'})
 
+        google_map = response.find('a', {'class': 'mod_epdetail__link-map'})
+        lat = None
+        lon = None
 
-    def parse(self, response):
-        hrefs = response.xpath("//h3/a[1]/@href")
-        for href in hrefs:
-            url = response.urljoin(href.extract())
-            yield scrapy.Request(url, callback = self.parse_contents)
-    
-        next_page_url = response.xpath("//li[@class='next']//a/@href").extract_first()
-        if next_page_url:
-            url = response.urljoin(next_page_url)
-            yield scrapy.Request(url)
-            
-    def parse_contents(self, response):
-        Ids = response.xpath("//div[@class='parameter__box']")
-        for Id in Ids:
-            nr = Id.xpath(
-                ".//li[4]/span/text()").extract()
-            
-        headlines = response.xpath("//div[@class='parameter__box']/div")
-        for headline in headlines:
-            hl = headline.xpath(
-                ".//h1/text()").extract_first()
-        
-        locations = response.xpath("//div[@class='parameter__box']")
-        for location in locations:
-            loc = location.xpath(
-                ".//li[1]/span/text()").extract()
-        
-        tasks = response.xpath("//div[2][@class='text-block']")
-        for task in tasks:
-            tk = task.xpath(
-                ".//p/text()").extract()
-       
-        times = response.xpath("//div[@class='parameter__box']")
-        for time in times:
-            tm = time.xpath(
-                ".//li[3]/span/text()").extract()
-            
-        organizations = response.xpath("//div[4][@class='text-block']")
-        for organization in organizations:
-            orga = organization.xpath(
-                ".//p[1]/text()").extract()
-            
-        yield {'id': nr,'title': hl, 'Location': loc, 'task': tk, 'timing': tm, 'organization': orga}
-            
+        if google_map:
+            try:
+                result = re.findall(r'q=(-?\d+\.\d+),(-?\d+\.\d+)', google_map['href'])
+                lat = float(result[0][0])
+                lon = float(result[0][1])
+            except (IndexError, TypeError, ValueError):
+                pass
 
-def runCrawler(name):
-    c = CrawlerProcess({
-        'USER_AGENT': 'HochschuleDarmstadt-MasterProjekt',
-        'FEED_FORMAT': 'csv',
-        'FEED_URI': name+'.csv',
-        'ROBOTSTXT_OBEY': True,
-        'HTTPCACHE_ENABLED': True
+        contact_raw = content.find('h2', text='Ansprechpartner*in und Entsendeorganisation').findNext('div').p.decode_contents()
+        contact_raw = re.sub(r'<br/?>', '\n', contact_raw)
+        contact_raw = contact_raw.replace('</br>', '')
+        contact_raw = contact_raw.replace('\n\n', '\n')
 
-    })
-    c.crawl(eval(name))
-    c.start() # the script will block here until the crawling is finished
+        parsed_object = {
+            'title': param_box.find("h1").decode_contents().strip() or None,
+            'link': url or None,
+            'organization': content.find('h2', text='Die Aufnahmeorganisation vor Ort').findNext('div').p.decode_contents().strip() or None,
+            'task': content.find('h2', text='Deine Aufgabe').findNext('div').p.decode_contents().strip() or None,
+            'location': param_box.find('li').find('span', {'class': 'parameter__value'}).decode_contents().strip() or None,
+            'contact': contact_raw.strip() or None,
+            'geo_location': {
+                'lat': lat,
+                'lon': lon,
+            },
+            'languages': param_box.find('li').findNext('li').find('span', {
+                'class': 'parameter__value'}).decode_contents().strip() or None,
+            'requirements': content.find('h2', text='Anforderungen an dich').findNext('div').p.decode_contents().strip() or None,
+            'source': self.base_url,
+            'image': None,
+        }
 
+        return parsed_object
 
-runCrawler('crawler')
+    #   Adds all URLs of detail pages, found on the search pages, for the crawl function to scrape
+    def add_urls(self):
+        import time
+
+        search_page_url = f'{self.base_url}/de/ep-ergebnis.html'
+        next_page_url = search_page_url
+
+        index = 1
+        while next_page_url:
+
+            response = self.soupify(next_page_url)
+
+            detail_link_tags = [x.find('a') for x in response.find_all('h3', {'class': 'result__headline'})]
+
+            if self.debug:
+                print(f'Fetched {len(detail_link_tags)} URLs from {next_page_url} [{index}]')
+
+            for link_tag in detail_link_tags:
+                current_link = self.base_url + '/' + link_tag['href']
+                if current_link in self.urls:
+                    self.add_error({
+                        'func': 'add_urls',
+                        'body': {
+                            'page_index': index,
+                            'search_page': next_page_url,
+                            'duplicate_link': current_link,
+                            'duplicate_index': self.urls.index(current_link)
+                        }
+                    })
+                else:
+                    self.urls.append(current_link)
+
+            next_page_url = response.find('li', {'class': 'next'}).find('a', {'class': 'next'})['href']
+            if next_page_url:
+                next_page_url = self.base_url + '/' + next_page_url
+
+            index += 1
+
+            time.sleep(self.delay)
